zoom/Backend_Files/requirements.txt
pyautogui
schedule
numpy
mysql.connector
opencv-python
